+++ src/skmultiflow/data/sine_generator.py
@@ -260,7 +260,7 @@ class SineGenerator(Stream):
                 data[j, 2] = group
 
         self.current_sample_x = data[:, :self.n_features]
-        self.current_sample_y = data[:, self.n_features:].flatten()
+        self.current_sample_y = data[:, self.n_features:].flatten().astype(int)
 
         return self.current_sample_x, self.current_sample_y
